Moved redis code to server side.
